+++ tests/image/test_ms_ssim.py
@@ -20,7 +20,7 @@ from pytorch_msssim import ms_ssim
 
 from tests.helpers import seed_all
 from tests.helpers.testers import NUM_BATCHES, MetricTester
-from torchmetrics.functional.image.ms_ssim import multiscale_structural_similarity_index_measure
+from torchmetrics.functional.image.ssim import multiscale_structural_similarity_index_measure
 from torchmetrics.image.ssim import MultiScaleStructuralSimilarityIndexMeasure
 
 seed_all(42)
