+++ torchmetrics/functional/__init__.py
@@ -37,9 +37,12 @@ from torchmetrics.functional.classification.roc import roc
 from torchmetrics.functional.classification.specificity import specificity
 from torchmetrics.functional.classification.stat_scores import stat_scores
 from torchmetrics.functional.image.gradients import image_gradients
-from torchmetrics.functional.image.ms_ssim import multiscale_structural_similarity_index_measure
 from torchmetrics.functional.image.psnr import psnr
-from torchmetrics.functional.image.ssim import ssim, structural_similarity_index_measure
+from torchmetrics.functional.image.ssim import (
+    multiscale_structural_similarity_index_measure,
+    ssim,
+    structural_similarity_index_measure,
+)
 from torchmetrics.functional.pairwise.cosine import pairwise_cosine_similarity
 from torchmetrics.functional.pairwise.euclidean import pairwise_euclidean_distance
 from torchmetrics.functional.pairwise.linear import pairwise_linear_similarity
